Switch to ffmpeg-python, move ClickLabel to utils
Using ffmpeg-python makes the commands slightly less error-prone, and made it
easier to improve the performance of the common use case of the .ass file
matching the aspect ratio of the output video. It should also make let me
finally implement that Override Background option much more easily, as well as
the Intro/Outro feature I'll soon be adding.