Fix logic issue for different aspect ratios
Subtitle should be letterboxed if there's an aspect ratio difference no matter what type of background is used. Previously it was only doing so with a picture or video background, meaning that if e.g. someone used CDG aspect ratio subtitle with 16:9 color background, it would have expanded to fill instead of letter boxing. There's no difference if the text is narrow and centered, but it did not behave correctly if any was left/right aligned or extended past the border of what should have been the CDG letterbox.