Outro basically works, added fade to black, removed intro/outro sound for now
Probably not the most efficient way to get it done... had some issues with the fade out not going to black when that was checked, and apparently if format is rgba, fade will use alpha if the color is black even if alpha=0. Very strange. But it also turns out that the fade generates video with alpha where appropriate even if the input format is not rgba.

Tested with the following config:
- image background
- image intro, fade from black
- image outro, fade to black

Need to test more possiblities

Also still need to add support for intro/outro overlap != length or disable that feature in the UI for now.