Initial fix for outro video length issue
In some cases, setting an outro video would briefly display the background again, and it seems to be due to the background length not exactly matching the outro length plus the start time. There may be two causes to this:

- Dealing in float times instead of frames causing rounding issues
- Containers with a slightly higher duration than the video stream

Will keep pursuing those, but in the meantime, repeating the last frame of the outro should be a good quick fix.