Experimental import of Doblon full timing lyrics export
If you drag in a Doblon lyrics export .txt file will full timing, it will be immediately converted to a .kbp file unless the filename already exists in that path. Then it will be treated like a normal kbp file that you can launch an editor for, or convert to .ass/video.

Note that there are some settings for it, but no UI for them yet. If you need to update them, on Windows they are in the registry. See HKEY_CURRENT_USER\SOFTWARE\ItMightBeKaraoke\kbp2video\doblontxt\