Use ms for durations (except in cases of using unmodified ffprobe strings)
This was kind of a mess before. The intro/outro prompt would return the durations as strings, and some math was done in floating point that didn't need to be. Now everything is milliseconds until it has to be converted to seconds (which is only the case for the fade filter), or if it can be passed directly as retrieved from ffprobe (in which case it's clearly spelled out with a _str suffix on the variable).

This would allow e.g. updating the intro/outro prompt to allow minutes for the fades without breaking anything if that were a desired feature and just makes things less brittle in general.