Accept files specified by command-line
Change command-line processing to the Qt way, and add a parameter for specifying files to import. This is exactly equivalent to dragging in the files after the app is opened.

Convert imported paths to absolute/normalized during import so they don't break things (less likely to be have caused an issue before from drop events, but now with arbitrary CLI arguments it's a requirement)