Match against single .ass file in drop like .kbp
This is hopefully the last bit of logic that was only applying to .kbp files but should have been applying to .ass as well.

If a single .ass file is dropped in with audio/bg, the filenames shouldn't need to fuzzy match because they were obviously intended for the one file.